pyproject.toml
[project]
name = "leah"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "pyyaml>=6.0.2",
]
